requirements.txt
google-generativeai>=0.8.0
mss>=9.0.0
Pillow>=10.0.0
